Extracted Sending and Computation elements from BackgroundDataSend
This refractor is about making the BackgroundDataSend and its new children DataSender & DataComputer more managable. This was a refractor to get the methods into more the smaller files with minimal changes internally.
