+++ client/android/Crowd-ML/app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
     <uses-permission android:name="android.permission.WAKE_LOCK" />
 
     <application
-        android:name=".firebase.CrowdMLApplication"
+        android:name=".CrowdMLApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -22,7 +22,7 @@
             </intent-filter>
         </activity>
         <!--<activity
-            android:name=".DataSend"
+            android:name=".DataSender"
             android:label="@string/app_name" />
         <activity
             android:name=".OfflineDataSend"
