+++ client/android/Crowd-ML/app/src/main/java/osu/crowd_ml/CrowdMLApplication.java
@@ -1,4 +1,4 @@
-package osu.crowd_ml.firebase;
+package osu.crowd_ml;
 
 import android.app.Application;
 import android.content.Context;
